+++ opencsp/common/lib/csp/test/test_Intersection.py
@@ -1,6 +1,8 @@
 """Unit test to test the MirrorPoint class"""
 
 import numpy as np
+import unittest
+
 
 from opencsp.common.lib.csp.LightPathEnsemble import LightPathEnsemble
 from opencsp.common.lib.csp.LightSourcePoint import LightSourcePoint
@@ -17,7 +19,7 @@ from opencsp.common.lib.geometry.Vxyz import Vxyz
 import opencsp.common.lib.csp.RayTrace as rt
 
 
-class TestIntersection:
+class TestIntersection(unittest.TestCase):
     """Test class for testing Intersection class. MirrorParametric and RayTrac must also be working."""
 
     def test_rays_intersect_properly(self):
@@ -108,9 +110,105 @@ class TestIntersection:
         # Test
         np.testing.assert_array_almost_equal(intersection.intersection_points._data, expected_points.data)
 
+    def test_bad_params_raises_error(self):
+        """Make sure that passing in bad parameters causes an exception"""
+        # too many values for plane
+        lines_points = Pxyz([0, 0, 0])
+        lines_vecs = Vxyz([0, 0, 0])
+        plane_point = Pxyz(np.array([[0, 0], [0, 0], [0, 0]]))
+        plane_normal = Pxyz(np.array([[0, 0], [0, 0], [0, 0]]))
+
+        with self.assertRaises(ValueError):
+            Intersection.plane_lines_intersection((lines_points, lines_vecs), (plane_point, plane_normal))
+
+        # unmatched dimensions for plane
+        with self.assertRaises(ValueError):
+            Intersection.plane_lines_intersection((lines_points, lines_vecs), (Pxyz([0, 0, 0]), plane_normal))
+        with self.assertRaises(ValueError):
+            Intersection.plane_lines_intersection((lines_points, lines_vecs), (plane_point, Pxyz([0, 0, 0])))
+
+        # unmatched dimensions for lines
+        with self.assertRaises(ValueError):
+            Intersection.plane_lines_intersection(
+                (Pxyz(np.array([[0, 0], [0, 0], [0, 0]])), lines_vecs), (plane_point, plane_normal)
+            )
+        with self.assertRaises(ValueError):
+            Intersection.plane_lines_intersection(
+                (lines_points, Pxyz(np.array([[0, 0], [0, 0], [0, 0]]))), (plane_point, plane_normal)
+            )
+
+    def test_parallel_line_plane(self):
+        """Raise an exception when the line and the plane are parallel"""
+        lines_points = Pxyz([0, 0, 0])
+        lines_vecs = Vxyz([0, 0, 1])
+        plane_point = Pxyz([1, 1, 1])
+        plane_normal = Uxyz([0, 1, 0])
+
+        with self.assertRaises(ValueError):
+            Intersection.plane_lines_intersection((lines_points, lines_vecs), (plane_point, plane_normal))
+
+    def test_parallel_coincident_line_plane(self):
+        """Raise an exception when the line and the plane are parallel, including the coincident edge case."""
+        lines_points = Pxyz([0, 0, 0])
+        lines_vecs = Vxyz([0, 0, 1])
+        plane_point = Pxyz([0, 0, 0])
+        plane_normal = Pxyz([0, 1, 0])
+
+        with self.assertRaises(ValueError):
+            Intersection.plane_lines_intersection((lines_points, lines_vecs), (plane_point, plane_normal))
+
+    def test_x_plane(self):
+        """Verify that we can find intersections for a plane with a normal == x-axis"""
+        lines_points = Pxyz([1, 0, 1])
+        lines_vecs = Vxyz([1, 0, 0])
+        plane_point = Pxyz([0, 0, 0])
+        plane_normal = Pxyz([1, 0, 0])
+
+        intersection_points = Intersection.plane_lines_intersection(
+            (lines_points, lines_vecs), (plane_point, plane_normal)
+        )
+        self.assertEqual(1.0, intersection_points.z)
+
+    def test_y_plane(self):
+        """Verify that we can find intersections for a plane with a normal == y-axis"""
+        lines_points = Pxyz([1, 1, 0])
+        lines_vecs = Vxyz([0, 1, 0])
+        plane_point = Pxyz([0, 0, 0])
+        plane_normal = Pxyz([0, 1, 0])
+
+        intersection_points = Intersection.plane_lines_intersection(
+            (lines_points, lines_vecs), (plane_point, plane_normal)
+        )
+        self.assertEqual(1.0, intersection_points.x)
+
+    def test_z_plane(self):
+        """Verify that we can find intersections for a plane with a normal == z-axis"""
+        lines_points = Pxyz([0, 1, 1])
+        lines_vecs = Vxyz([0, 0, 1])
+        plane_point = Pxyz([0, 0, 0])
+        plane_normal = Pxyz([0, 0, 1])
+
+        intersection_points = Intersection.plane_lines_intersection(
+            (lines_points, lines_vecs), (plane_point, plane_normal)
+        )
+        self.assertEqual(1.0, intersection_points.y)
+
+    def test_multiple_lines(self):
+        """Test with two lines and with three lines. We should get the same number of intersection points as the number of lines."""
+        lines_points = Pxyz([[-1, 0, 1], [0, 0, 0], [1, 1, 1]])
+        lines_vecs = Vxyz([[1, 0, 0], [0, 1, 0], [-1, -1, -1]])
+        plane_point = Pxyz([0, 0, 0])
+        plane_normal = Pxyz([0, 0, 1])
+
+        intersection_points = Intersection.plane_lines_intersection(
+            (lines_points, lines_vecs), (plane_point, plane_normal)
+        )
+        self.assertEqual(3, intersection_points.data.shape[1])
+
 
 if __name__ == '__main__':
     Test = TestIntersection()
     Test.test_rays_intersect_properly()
     Test.test_parallel_reflection()
     Test.test_converging_reflection()
+    unittest.main()
